:sparkles: Start the key map
We're going to start with using Enter to open the URL in a browser. For now
though I'm out of time so I'm pushing this up to finish it off later.

Note to self: I need to write a function that, given the hash of a pin (as
that's the ID we're using for the table), it finds the pin that's come down
in the structure from Pinboard.