:sparkles: Update the local list when we make changes
The visible list doesn't get updated (for now anyway), but if the user adds
or edits a pin and then hist "g" in the list it'll show what they did. The
local changes are "good enough" for display purposes and when the user
eventually pulls from the server all will be better.