Call auth-source-search to require a password
Updates call to auth-source-search use its API correctly, in order to
return only entries that provide both the :host (aka, the "machine")
and a :secret (aka, a "password") for accessing the machine. This
updated call corrects an error in the existing call to this function,
which seems to require that returned entries have a secret, but in
fact does not require this because of two typos -- using :requires
instead of :require, and using 'secret instead of :secret.

Longer discussion:

The docstring and source code for auth-source-search indicate that it
can take an optional keyword :require, in order to pass in a list of
required auth search query keys. These query keys should be keyword
symbols. Some such auth search query keys are specified in the "Help
for users" info page for auth, and in lines 2291-2294 of
auth-source.el.gz version 23.1. They are :host, :user, :port, and
:secret.

This commit updates the function call to use this :require keyword,
and to pass in the actually required query key, :secret.

If the previous commit was calling the API incorrectly, why did it
work? The previous commit worked by accident, because the undefined
keyword :requires was ignored. Thus no property besides :machine was
in fact being required. The function was returning all properties by
default, which includes :secret but only as long as it was present.

This can be confirmed by noting that the old call to
auth-source-search returns results even if the .authinfo file defines
a "machine" but not a "password".

Changing :requires to :require causes the call to return nothing,
because then the problem is that 'secret is being required but it is
given as the symbol 'secret and a valid query key must be a keyword
symbol like :secret. The solution therefore is to change :requires to
:require and also 'secret to :secret.

What's the benefit of this commit? First, calling the auth API
correctly should allow pinboard to throw correct errors if the
authinfo database does not have a defined password. Second, it should
make enhancements easier in the future, in case the auth-source-search
stops returning all properties by default, or in case a developer
wants pinboard.el to require new properties and follows the plan of
existing code.